km_vernacularai/requirements.txt
Django==3.0.8
djangorestframework
pylint
marshmallow
